+++ test-suite/bugs/closed/bug_5233.v
@@ -1,2 +1,2 @@
 (* Implicit arguments on type were missing for recursive records *)
-Inductive foo {A : Type} : Type := { Foo : foo }.
+Inductive foo {A : Type} (a : A) : Type := { Foo : foo a }.